+++ packages/plugin-e2e/src/fixtures/commands/createDataSource.ts
@@ -1,7 +1,6 @@
 import { v4 as uuidv4 } from 'uuid';
 import { APIRequestContext, expect, TestFixture } from '@playwright/test';
-import { PluginFixture, PluginOptions } from '../../api';
-import { CreateDataSourceArgs, DataSourceSettings } from '../../types';
+import { CreateDataSourceArgs, DataSourceSettings, PluginFixture, PluginOptions } from '../../types';
 import { PlaywrightCombinedArgs } from '../types';
 
 type CreateDataSourceViaAPIFixture = TestFixture<
