+++ packages/plugin-e2e/src/fixtures/explorePage.ts
@@ -1,5 +1,5 @@
 import { TestFixture } from '@playwright/test';
-import { PluginFixture, PluginOptions } from '../api';
+import { PluginFixture, PluginOptions } from '../types';
 import { ExplorePage } from '../models/pages/ExplorePage';
 import { PlaywrightCombinedArgs } from './types';
