+++ packages/plugin-e2e/src/fixtures/page.ts
@@ -1,5 +1,5 @@
 import { TestFixture, Page } from '@playwright/test';
-import { PluginFixture, PluginOptions } from '../api';
+import { PluginFixture, PluginOptions } from '../types';
 import { PlaywrightCombinedArgs } from './types';
 import { overrideFeatureToggles } from './scripts/overrideFeatureToggles';
